Broken cursor bugfix for 64 bit systems (#49)
* Proper 32 bit types assigned

* The size dependent conversions are not necessary if input type is 32 bit in every case

* Unused variable

* Another type problem 32/64 fixed

* Prototype fixed

* Pointer cast missing

* * Multiple conversions from (unsigned long*) to proper (uint32_t*).
* Bugfix for XFixesCursorImage type problem. Structure contains (unsigned long*), but protocol says the pixels are just 32 bit.
  This is a problem for 64 bit systems where (unsigned long) is 64 bits wide.

+++ src/cursor.h
@@ -47,7 +47,6 @@ extern int alpha_blend;
 extern int alt_arrow;
 extern int alt_arrow_max;
 
-
 extern void first_cursor(void);
 extern void setup_cursors_and_push(void);
 extern void initialize_xfixes(void);
@@ -64,7 +63,7 @@ extern void set_no_cursor(void);
 extern void set_warrow_cursor(void);
 extern int set_cursor(int x, int y, int which);
 extern int check_x11_pointer(void);
-extern int store_cursor(int serial, unsigned long *data, int w, int h, int cbpp, int xhot, int yhot);
+extern int store_cursor(int serial, uint32_t *data, int w, int h, int cbpp, int xhot, int yhot);
 extern unsigned long get_cursor_serial(int mode);
 extern rfbCursorPtr pixels2curs(uint32_t *pixels, int w, int h, int xhot, int yhot, int Bpp);
 void save_under_cursor_buffer(rfbClientPtr cl);
